move /renderer to /client/renderer
