+++ src/main/java/gregicadditions/client/renderer/onlinepictexture/PictureTexture.java
@@ -1,4 +1,4 @@
-package gregicadditions.renderer.onlinepictexture;
+package gregicadditions.client.renderer.onlinepictexture;
 
 import net.minecraft.client.renderer.GlStateManager;
 import net.minecraft.client.renderer.OpenGlHelper;
