+++ src/main/java/gregicadditions/item/behaviors/monitorPlugin/AdvancedMonitorPluginBehavior.java
@@ -8,8 +8,8 @@ import codechicken.lib.vec.Translation;
 import codechicken.lib.vec.Vector3;
 import gregicadditions.machines.multi.centralmonitor.MetaTileEntityCentralMonitor;
 import gregicadditions.machines.multi.centralmonitor.MetaTileEntityMonitorScreen;
-import gregicadditions.renderer.PluginWorldSceneRenderer;
-import gregicadditions.renderer.RenderHelper;
+import gregicadditions.client.renderer.PluginWorldSceneRenderer;
+import gregicadditions.client.renderer.RenderHelper;
 import gregicadditions.utils.BlockPatternChecker;
 import gregicadditions.utils.GALog;
 import gregicadditions.utils.Tuple;
