+++ src/main/java/gregicadditions/machines/multi/centralmonitor/MetaTileEntityCentralMonitor.java
@@ -5,8 +5,8 @@ import codechicken.lib.render.pipeline.IVertexOperation;
 import codechicken.lib.vec.Matrix4;
 import gregicadditions.covers.CoverDigitalInterface;
 import gregicadditions.item.GAMetaBlocks;
-import gregicadditions.renderer.RenderHelper;
 import gregicadditions.utils.BlockPatternChecker;
+import gregicadditions.client.renderer.RenderHelper;
 import gregicadditions.utils.Tuple;
 import gregicadditions.widgets.monitor.WidgetScreenGrid;
 import gregtech.api.capability.GregtechCapabilities;
@@ -17,7 +17,6 @@ import gregtech.api.gui.ModularUI;
 import gregtech.api.gui.Widget;
 import gregtech.api.gui.widgets.AdvancedTextWidget;
 import gregtech.api.gui.widgets.WidgetGroup;
-import gregtech.api.metatileentity.IFastRenderMetaTileEntity;
 import gregtech.api.metatileentity.IRenderMetaTileEntity;
 import gregtech.api.metatileentity.MetaTileEntity;
 import gregtech.api.metatileentity.MetaTileEntityHolder;
