+++ src/main/java/gregicadditions/machines/multi/centralmonitor/MetaTileEntityMonitorScreen.java
@@ -6,8 +6,8 @@ import gregicadditions.covers.CoverDigitalInterface;
 import gregicadditions.item.behaviors.monitorPlugin.ProxyHolderPluginBehavior;
 import gregicadditions.item.behaviors.monitorPlugin.MonitorPluginBaseBehavior;
 import gregicadditions.machines.GATileEntities;
-import gregicadditions.renderer.RenderHelper;
 import gregicadditions.utils.BlockPatternChecker;
+import gregicadditions.client.renderer.RenderHelper;
 import gregicadditions.utils.Tuple;
 import gregicadditions.widgets.WidgetARGB;
 import gregicadditions.widgets.monitor.WidgetMonitorScreen;
